apps/hermes/requirements.txt
python-telegram-bot==21.9
httpx==0.28.1
redis[hiredis]==5.2.1
